\name{get_connection}
\alias{get_connection}
\title{Get a Connection with WTT-01 loaded}
\usage{
  get_connection(dbdir = ":memory:")
}
\arguments{
  \item{dbdir}{The directory where the database file is located.}
}
\description{
  This function returns a connection to a duckdb database file located in the specified directory with WTT-01 loaded.
}
